// src/app/shared/enums/firebase.enum.ts
export enum ECollectionFirebase {
  PESSOAS = 'pessoas',
  INDICANTES = 'indicantes',
  ESCOLAS = 'escolas',
  SECRETARIAS = 'secretarias',
  ORDEM_COMPRA = 'ordem_compra',
  USUARIOS = 'usuarios',
  PRODUTOS = 'produtos',
  FORNECEDORES = 'fornecedores',
  NOTAS_FISCAIS = 'notas_fiscais',
  COLABORADORES = 'colaboradores',
}